refactor FunnelsCue
