+++ frontend/src/scenes/insights/InsightsNav.tsx
@@ -28,7 +28,7 @@ interface Tab {
 }
 
 export function InsightsNav(): JSX.Element {
-    const { activeView, insightProps, createInsightUrl } = useValues(insightLogic)
+    const { activeView, createInsightUrl } = useValues(insightLogic)
     const { setActiveView } = useActions(insightLogic)
     const { featureFlags } = useValues(featureFlagLogic)
     const funnelTab = useRef<HTMLSpanElement>(null)
@@ -86,7 +86,6 @@ export function InsightsNav(): JSX.Element {
     return (
         <>
             <FunnelsCue
-                props={insightProps}
                 tooltipPosition={
                     // 1.5x because it's 2 tabs (trends & funnels) + margin between tabs
                     funnelTab?.current ? funnelTab.current.getBoundingClientRect().width * 1.5 + 16 : undefined
